Save Slot System 2
Added more functionality to save slots, including the ability to delete and rename them, as well as sending the remote function to start the game.

Deleting is still buggy when the player renames, then deletes the save. It seems like the profile is not being released even though it is??