Room Events 1
- Added a system that determines what "event" should trigger when the player walks in the room.

For now, it is basic stuff like "curio" and  "encounter." The specifics of the event will be dependent on the room, whenever I get to adding that.

There is a chance for it to be empty, as indicated by the event "", in which the event trigger brick will be destroyed.